+++ CHANGELOG.md
@@ -101,4 +101,7 @@ fine for performance). Right now only Punish and Forgive are required to be logg
 <h4>0.1.9.1 (20-MAY-2013)</h4>
 <b>Changes</b> <br/>
    * Added banning admin name to all ban logs.<br/>
+<h4>0.1.9.2 (20-MAY-2013)</h4>
+<b>Changes</b> <br/>
+   * Added player mute system.<br/>
 </blockquote>